Clean quality boundaries
