--- packages/extension/src/extension/pipeline/plugins/connectionCounts.ts
@@ -1,3 +1,0 @@
-export {
-  countWorkspaceIndexPluginConnections as countPluginConnections,
-} from '@codegraphy-dev/core';
