--- packages/extension/src/extension/pipeline/plugins/extensions.ts
@@ -1,4 +1,0 @@
-export {
-  getWorkspaceIndexPluginMatchingFiles as getPluginMatchingFiles,
-  supportsWorkspaceIndexPluginExtension as supportsExtension,
-} from '@codegraphy-dev/core';
