--- packages/extension/src/extension/pipeline/plugins/statusBuilder.ts
@@ -1,4 +1,0 @@
-export {
-  buildWorkspaceIndexPluginStatuses as buildWorkspacePluginStatuses,
-  type WorkspaceIndexPluginStatusOptions as IWorkspacePluginStatusOptions,
-} from '@codegraphy-dev/core';
