--- packages/extension/src/extension/pipeline/plugins/statusRecords.ts
@@ -1,6 +1,0 @@
-export {
-  buildRegisteredWorkspaceIndexPluginStatus as buildRegisteredPluginStatus,
-  buildUnregisteredInstalledWorkspaceIndexPluginStatus as buildUnregisteredInstalledPluginStatus,
-  getRegisteredWorkspaceIndexPluginPackageNames as getRegisteredPackageNames,
-  isUserFacingWorkspaceIndexPlugin as isUserFacingPlugin,
-} from '@codegraphy-dev/core';
