--- packages/extension/src/webview/components/graph/runtime/physics/nodeLookup.ts
@@ -1,5 +1,0 @@
-import type { FGNode } from '../../model/build';
-
-export function createNodeMap(nodes: readonly FGNode[]): Map<string, FGNode> {
-	return new Map(nodes.map(node => [node.id, node]));
-}
