+++ packages/extension/src/webview/pluginHost/disposable.ts
@@ -1,0 +1,7 @@
+export interface WebviewDisposable {
+  dispose(): void;
+}
+
+export function toWebviewDisposable(dispose: () => void): WebviewDisposable {
+  return { dispose };
+}
